norminette 1ver done
