+++ srcs/builtin_echo.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtin_echo.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seseo <seseo@student.42.fr>                +#+  +:+       +#+        */
+/*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/28 00:11:28 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/09 15:27:37 by seseo            ###   ########.fr       */
+/*   Updated: 2022/07/09 19:20:09 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
