+++ srcs/heredoc.c
@@ -6,12 +6,11 @@
 /*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/27 23:19:21 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/09 17:24:22 by wchae            ###   ########.fr       */
+/*   Updated: 2022/07/09 19:31:03 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
 
 static void	heredoc_child(char *limiter, int *fd)
 {
