+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/11 16:28:59 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/09 19:18:03 by wchae            ###   ########.fr       */
+/*   Updated: 2022/07/09 19:30:23 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	check_token(t_token	*token)
 {
-	int i;
+	int	i;
 
 	i = 0;
 	while (token)
@@ -40,67 +40,10 @@ int	check_token(t_token	*token)
 	return (TRUE);
 }
 
-
-
-char	*skip_space(char *str)
+void	backup_fd(int backup_io[2])
 {
-	while (*str)
-	{
-		if (*str != ' ')
-			break ;
-		str++;
-	}
-	return (str);
-}
-
-char	**split_skip_quote(char *str)
-{
-	t_buffer	*buf;
-	t_list		*tmp_lst;
-	char		**strs;
-
-	tmp_lst = NULL;
-	buf = create_buf();
-	while (*str)
-	{
-		if (is_quote(*str))
-			str = skip_quote_2(buf, str, is_quote(*str));
-		else if (*str != ' ')
-			add_char(buf, *str);
-		else
-		{
-			str = skip_space(str);
-			ft_lstadd_back(&tmp_lst, ft_lstnew(put_str(buf)));
-			if (*str == 0)
-				break ;
-		}
-		str++;
-	}
-	strs = lst_to_strs(tmp_lst);
-	ft_lstclear(&tmp_lst, free);
-	del_buf(buf);
-	return (strs);
-}
-
-char	**split_cmd(t_list *cmd)
-{
-	char	**exe;
-	int		size;
-	int		i;
-
-	size = ft_lstsize(cmd);
-	exe = (char **)malloc((size + 1) * sizeof(char *));
-	if (!exe)
-		return (NULL);
-	i = 0;
-	while (i < size)
-	{
-		exe[i] = cmd->data;
-		cmd = cmd->next;
-		i++;
-	}
-	exe[i] = NULL;
-	return (exe);
+	backup_io[0] = dup(STDIN_FILENO);
+	backup_io[1] = dup(STDOUT_FILENO);
 }
 
 t_token	*parse_input(char *input, t_env *env)
@@ -122,13 +65,7 @@ t_token	*parse_input(char *input, t_env *env)
 	return (token);
 }
 
-void	backup_fd(int backup_io[2])
-{
-	backup_io[0] = dup(STDIN_FILENO);
-	backup_io[1] = dup(STDOUT_FILENO);
-}
-
-int main(void)
+int	main(void)
 {
 	t_token	*tokens;
 	t_set	set;
